combining the two addresses components into one
