+++ src/components/AddressesList.css
@@ -1,3 +1,33 @@
+.rounded-border {
+    border-radius: .75rem;
+    border: 2px solid white;
+    padding: 1rem;
+    background-color: white;
+}
+
+.column-flex {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.vertical-align {
+    display: flex;
+    align-items: center;
+}
+
+.icon {
+    margin: 0 .5rem;
+}
+
+.addresses-box {
+    width: 17vw;
+}
+
+.validation-error {
+    color: red;
+}
+
 .list-margin {
     margin-top: .5vh;
 }
