+++ src/pages/Origins.tsx
@@ -1,5 +1,4 @@
 import "./Origins.css"
-import AddressesInput from "../components/AddressesInput";
 import StatusBar from "../components/StatusBar";
 import AddressesList from "../components/AddressesList";
 import {Button} from "@mui/material";
@@ -8,8 +7,7 @@ export default function Origins() {
     return (
         <div className="center-items">
             <StatusBar activeStep={0} />
-            <AddressesInput typeOfAddress="origin" />
-            <AddressesList />
+            <AddressesList typeOfAddress="origin" />
             <Button variant="contained" className="next-step">Next Step</Button>
         </div>
     )
